pyproject.toml
[project]
name = "events_face"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [ "ruff>=0.9.10", "asgiref==3.8.1", "certifi==2025.1.31", "charset-normalizer==3.4.1", "Django==5.1.7", "django-filter==25.1", "djangorestframework==3.15.2", "djangorestframework_simplejwt==5.5.0", "docker==7.1.0", "idna==3.10", "PyJWT==2.9.0", "pywin32==310", "requests==2.32.3", "ruff==0.11.0", "sqlparse==0.5.3", "toml==0.10.2", "tzdata==2025.1", "urllib3==2.3.0", "uv==0.6.7", "amqp==5.3.1", "asgiref==3.8.1", "billiard==4.2.1", "celery==5.4.0", "certifi==2025.1.31", "charset-normalizer==3.4.1", "click==8.1.8", "click-didyoumean==0.3.1", "click-plugins==1.1.1", "click-repl==0.3.0", "colorama==0.4.6", "Django==5.1.7", "django-filter==25.1", "djangorestframework==3.15.2", "djangorestframework_simplejwt==5.5.0", "docker==7.1.0", "idna==3.10", "kombu==5.5.0", "prompt_toolkit==3.0.50", "PyJWT==2.9.0", "python-dateutil==2.9.0.post0", "pywin32==310", "redis==5.2.1", "requests==2.32.3", "ruff==0.11.0", "six==1.17.0", "sqlparse==0.5.3", "toml==0.10.2", "tzdata==2025.1", "urllib3==2.3.0", "uv==0.6.7", "vine==5.1.0", "wcwidth==0.2.13",]

[tool.ruff.lint]
extend-select = [ "I",]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
